Cleaning up mac compiler warnings
[git-p4: depot-paths = "//depot/Golf/": change = 1149]

+++ code/engine/SoundEngine.cpp
@@ -813,7 +813,7 @@ class BackgroundTrackMgr
 			OSStatus result = AudioQueuePrime(mQueue, 1, NULL);	
 			if (result)
 			{
-				printf("Error priming queue: %d\n", result);
+				printf("Error priming queue: %d\n", (int)result);
 				return result;
 			}
 			return AudioQueueStart(mQueue, NULL);
